Split Invite types
