+++ packages/types/src/discord/auditLogs.ts
@@ -1,20 +1,13 @@
 /** Types for: https://discord.com/developers/docs/resources/audit-log */
 
-import type {
-  DiscordInvite,
-  DiscordInviteMetadata,
-  DiscordRole,
-  DiscordStageInstance,
-  DiscordSticker,
-  DiscordUser,
-  DiscordWebhook,
-} from '../discord.js'
+import type { DiscordRole, DiscordStageInstance, DiscordSticker, DiscordUser, DiscordWebhook } from '../discord.js'
 import type { DiscordAutoModerationRule } from './autoModeration.js'
 import type { DiscordChannel, DiscordOverwrite, DiscordThreadMetadata } from './channels.js'
 import type { DiscordEmoji } from './emojis.js'
 import type { DiscordScheduledEvent } from './guildScheduledEvents.js'
 import type { DiscordGuild, DiscordGuildOnboarding, DiscordGuildOnboardingPrompt, DiscordIntegration, DiscordMember } from './guilds.js'
 import type { DiscordApplicationCommand, DiscordApplicationCommandPermissions } from './interactions.js'
+import type { DiscordInvite, DiscordInviteMetadata } from './invites.js'
 
 /** https://discord.com/developers/docs/resources/audit-log#audit-log-object-audit-log-structure */
 export interface DiscordAuditLog {
